+++ packages/opentelemetry-sdk-trace-base/src/platform/browser/index.ts
@@ -15,3 +15,4 @@
  */
 
 export * from './export/BatchSpanProcessor';
+export * from './RandomIdGenerator';